Add the common ontology patterns library
The patterns included in this library are part of the emerging Object Management Group (OMG) Multiple Vocabulary Facility (MVF) specification.  Adoption is anticipated in Q2 2022.  These patterns, some of which are embedded in the OMG's Languages, Countries, and Codes (LCC) standard and in FIBO at the moment, are stable and broadly useful across domains.  The patterns are largely derived from a combination of ISO standards, including but not limited to ISO 1087, ISO 11179, ISO 8601, and others.